# Fix potential deadlock when generating a crash report