--- a/Source/KSCrash/Source/KSCrash/Recording/BSG_KSCrashReport.c
+++ b/Source/KSCrash/Source/KSCrash/Recording/BSG_KSCrashReport.c
@@ -1085,21 +1085,23 @@ int bsg_kscrw_i_threadIndex(const thread_t thread) {
  *
  * @param key The object key, if needed.
  *
- * @param info Cached info about the binary image.
+ * @param index Cached info about the binary image.
  */
 void bsg_kscrw_i_writeBinaryImage(const BSG_KSCrashReportWriter *const writer,
                                   const char *const key,
-                                  const BSG_Mach_Binary_Image_Info *info)
+                                  const uint32_t index)
 {
+    BSG_Mach_Binary_Image_Info info = *bsg_dyld_get_image_info(index);
+    
     writer->beginObject(writer, key);
     {
-        writer->addUIntegerElement(writer, BSG_KSCrashField_ImageAddress, (uintptr_t)info->header);
-        writer->addUIntegerElement(writer, BSG_KSCrashField_ImageVmAddress,          info->imageVmAddr);
-        writer->addUIntegerElement(writer, BSG_KSCrashField_ImageSize,               info->imageSize);
-        writer->addStringElement(writer, BSG_KSCrashField_Name,                      info->name);
-        writer->addUUIDElement(writer, BSG_KSCrashField_UUID,                        info->uuid);
-        writer->addIntegerElement(writer, BSG_KSCrashField_CPUType,                  info->header->cputype);
-        writer->addIntegerElement(writer, BSG_KSCrashField_CPUSubType,               info->header->cpusubtype);
+        writer->addUIntegerElement(writer, BSG_KSCrashField_ImageAddress, (uintptr_t)info.header);
+        writer->addUIntegerElement(writer, BSG_KSCrashField_ImageVmAddress,          info.imageVmAddr);
+        writer->addUIntegerElement(writer, BSG_KSCrashField_ImageSize,               info.imageSize);
+        writer->addStringElement(writer, BSG_KSCrashField_Name,                      info.name);
+        writer->addUUIDElement(writer, BSG_KSCrashField_UUID,                        info.uuid);
+        writer->addIntegerElement(writer, BSG_KSCrashField_CPUType,                  info.header->cputype);
+        writer->addIntegerElement(writer, BSG_KSCrashField_CPUSubType,               info.header->cpusubtype);
     }
     writer->endContainer(writer);
 }
@@ -1114,13 +1116,12 @@ void bsg_kscrw_i_writeBinaryImages(const BSG_KSCrashReportWriter *const writer,
                                    const char *const key)
 {
     const uint32_t imageCount = (uint32_t)bsg_dyld_image_count();
-    
+
     writer->beginArray(writer, key);
     {
-        for (uint32_t i = 0; i < imageCount; i++) {
+        for (uint32_t iImg = 0; iImg < imageCount; iImg++) {
             // Threads are suspended at this point; no need to synchronise/lock
-            BSG_Mach_Binary_Image_Info *info = bsg_dyld_get_image_info(i);
-            bsg_kscrw_i_writeBinaryImage(writer, NULL, info);
+            bsg_kscrw_i_writeBinaryImage(writer, NULL, iImg);
         }
     }
     writer->endContainer(writer);