… .github/PULL_REQUEST_TEMPLATE.md
… CHANGELOG.md
… Source/KSCrash/Source/KSCrash/Recording/Tools/BSG_KSMachHeaders.h
… Tests/KSCrash/KSMachHeader_Tests.m
… iOS/Bugsnag.xcodeproj/project.pbxproj
… tvOS/Bugsnag.xcodeproj/project.pbxproj